Migrated image-pixels to pngjs
pngjs seems to be more widely used so I therefore decided to move from image-pixels, this was surprisingly easy to do.

I was already using pngjs to generate the preview images of the maps (i.e. #8) so I thought it might help keep down the amount of dependencies by using just pngjs instead.